Add gym mismatch check

+++ app/routes/employee_routes.py
@@ -2,7 +2,8 @@ from flask import Blueprint, request, jsonify
 from app.models import Employee, GymClass
 from app import db
 import logging
-from utils import role_required
+from utils import role_required, check_gym_mismatch
+from flask_jwt_extended import get_jwt
 
 employee_routes = Blueprint('employee_routes', __name__)
 
@@ -24,8 +25,8 @@ def update_employee(employee_id):
         logging.warning(f"Employee with ID {employee_id} does not exist")
         return jsonify({"msg": "Employee does not exist"}), 404
 
-    allowed_fields = {'password', 'gym_id', 'first_name', 'last_name', 'role'}
-
+    allowed_fields = {'password', 'first_name', 'last_name', 'role'}
+    
     # Check if the role is being changed from 'coach' to another role
     if 'role' in data and data['role'] != employee.role and employee.role == 'coach':
         try:
@@ -70,12 +71,20 @@ def update_employee(employee_id):
 @employee_routes.route('/delete_employee/<int:employee_id>', methods=['DELETE'])
 @role_required(["manager"])
 def delete_employee(employee_id):
+
     try:
         employee = Employee.query.get(employee_id)
         if not employee:
             logging.warning(f"Employee with ID {employee_id} does not exist")
             return jsonify({"msg": "Employee does not exist"}), 404
+        
+        jwt_payload = get_jwt()
+        user_gym_id = jwt_payload.get('gym_id')
 
+        if user_gym_id != employee.gym_id:
+            logging.warning("You are not authorized to modify this gym")
+            return jsonify({"msg": "You are not authorized to modify this gym"}), 403
+        
         db.session.delete(employee)
         db.session.commit()
         logging.info(f"Employee deleted successfully: ID {employee_id}")
@@ -93,7 +102,7 @@ def get_employee(employee_id):
     if not employee:
         logging.warning(f"Employee with ID {employee_id} does not exist")
         return jsonify({"msg": "Employee does not exist"}), 404
-
+    
     result = {
         "employee_id": employee.employee_id,
         "gym_id": employee.gym_id,
